settings: move Export / Import Backup from Workouts toolbar into Settings
The Workouts toolbar mixed day-navigation (Back / Advance) with data
controls (Export / Import Backup / Reset), forcing the data buttons to
be always-visible at the bottom of the workout view. The data controls
belong with other configuration in Settings.

This commit adds a new Data section to Settings hosting Export and
Import Backup, and removes those controls from the Workouts toolbar.
The program-import flow (empty-state CTA) is unchanged.

Architecture note: App.tsx conditionally mounts each tab, so when the
user switches back to Workouts after a backup import in Settings, the
Workouts effect re-runs loadState / loadLog / loadProgram and picks up
the fresh data from IDB. No cross-tab event system needed — Settings
just writes through the existing exportImport helpers.

Reset stays in the Workouts toolbar for now; it's destructive and
groups naturally with backup data so it may also belong in Settings,
left as a separate decision.
